+++ resources/views/errors/403.blade.php
@@ -1,0 +1,7 @@
+@extends('errors.minimal')
+
+@section('code', '403')
+@section('title', 'Access denied')
+@section('message')
+    {{ $exception->getMessage() ?: 'You do not have permission to view this page.' }}
+@endsection